+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- `tactics2d.renderer` module
+
 ### Changed
 
 - Split `tactics2d.math` into `tactics2d.interpolator` and `tactics2d.geometry`.
@@ -21,6 +23,9 @@
 
 ### Removed
 
+- `tactics2d.sensor` module
+- `tactics2d.traffic.scenario_display`
+
 ### TODO
 
 - `tactics2d.interpolator.cubic_spline`: Improve efficiency of CubicSpline by ThomasSolver.
@@ -29,7 +34,6 @@
 - `tactics2d.dataset_parser.womd_proto`: Add compatibility to protobuf 3.x.x and 4.x.x.
 - `tactics2d.map.parser.OSMParser`: Handle the tag `highway` in `load_way` for the original [OSM label style](https://wiki.openstreetmap.org/wiki/Key:lanes).
 - `tactics2d.dataset_parser`: Improve the efficiency.
-- Complete the CLI commands.
 
 ## [0.1.8] - 2025-05-22